aider: Remove call to `processCSS` in `build.js` to prevent the code from `build-css.js` being executed twice.
# Aider chat conversation:

USER: with the current code, we have the problem, that the code from build-css.js gets executed twice. what can we do about that?

ASSISTANT: The issue of the code from `build-css.js` being executed twice is likely due to the fact that the `processCSS` function is being called both when the module is imported and when the module is run directly.

To resolve this, we can remove the call to `processCSS` in `build.js`. This way, `processCSS` will only be called when `build-css.js` is run directly, and not when it's imported as a module.

Here's the change in the form of an *edit block*:

```javascript
scripts/build.js
<<<<<<< ORIGINAL
const { processCSS } = require('./build-css.js');

// Start build
(async function doBuild() {
  await displayAsciiArt('Simple Rick´s');
  console.log('Vanilla PWA Builder\n')
  runPug();
  copyAll();
  processCSS();
  purgeCSS();
  minifyJS();
  minifyHTML();
  console.log('Build finished!');
  startBrowserSyncBuild();
})();
=======
const { processCSS } = require('./build-css.js');

// Start build
(async function doBuild() {
  await displayAsciiArt('Simple Rick´s');
  console.log('Vanilla PWA Builder\n')
  runPug();
  copyAll();
  // processCSS(); // Commented out this line
  purgeCSS();
  minifyJS();
  minifyHTML();
  console.log('Build finished!');
  startBrowserSyncBuild();
})();
>>>>>>> UPDATED
```

With this change, `processCSS` will only be called when `build-css.js` is run directly, and not when it's imported as a module. This should prevent the code from being executed twice.