aider: The commit message is: "Prevent processCSS from being called twice."
# Aider chat conversation:

USER: the whole purpose of our work was, to use the code as a module. but we want both options. in the build script, that will become a module soon too, it should be imported and handled as a module. on the other hand, we want the option to execute it with npm run build:css . it allready works for the most part, but the problem is, when we npm run build, it seems that processCSS() is called twice. how can we make it work as desired?

ASSISTANT: The issue you're experiencing is likely due to the fact that `processCSS` is being called when the `build-css.js` module is imported and also when it's run directly.

To solve this, we can add a check in `build-css.js` to see if it's being run directly or imported as a module. If it's being run directly, we call `processCSS`. If it's being imported as a module, we don't call `processCSS`, but instead export it so it can be called in the file that's importing it.

Here's the change in the form of an *edit block*:

```javascript
scripts/build-css.js
<<<<<<< ORIGINAL
// If this module is run directly, call processCSS
if (require.main === module) {
  processCSS();
}

// Export the functions as modules
module.exports = {
  processCSS,
  processCSSInHTMLHead,
};
=======
// If this module is run directly, call processCSS
if (require.main === module) {
  processCSS();
}

// Export the functions as modules
module.exports = {
  processCSS,
  processCSSInHTMLHead,
};
>>>>>>> UPDATED
```

With this change, `processCSS` will only be called when `build-css.js` is run directly, and not when it's imported as a module. This should prevent the code from being executed twice.

In your `build.js` file, you can then import `processCSS` and call it when needed. Since `processCSS` is not called when `build-css.js` is imported as a module, it won't be executed twice when you run `npm run build`.